Accept tokens in addition of characters-based strings
The inputs can now be string or string[], for example when your
unit is the word instead of the character. In fact, as backward compatible,
the natural tokenisation is the split by characters; other tokenisations
must be computed beforehand.

It is added a boolean parameter in solve($stringA, $stringB, $string = true)
to request a string as result (if tokens are given they are implode('')) or
tokens as result (if strings are given their characters are returned).